pyproject.toml
[project]
name = "agentic-rag-architectures"
version = "0.1.0"
description = "Agentic RAG Architectures"
requires-python = ">=3.12"
dependencies = [
  "adjustText==1.3.0",
  "instructor==1.13.0",
  "matplotlib==3.10.7",
  "nbformat==5.10.4",
  "numpy==2.3.5",
  "openai==2.8.1",
  "pandas==2.3.3",
  "plotly==6.5.0",
  "pydantic==2.12.4",
  "python-dotenv==1.2.1",
  "pyyaml==6.0.2",
  "pytest==8.4.1",
  "ruff==0.14.6",
  "sentence-transformers==5.1.2",
  "seaborn==0.13.2",
  "pyarrow>=23.0.0",
  "fastparquet>=2025.12.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["."]
